refactor: a little bit refactor
