+++ backend/src/FileService/FileService.Core/Files/FileKey/IFileKeyGenerator.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace FileService.Core.Files.FileKey;
+﻿namespace FileService.Core.Files.FileKey;
 
 public interface IFileKeyGenerator
 {
