+++ backend/src/FileService/FileService.Infrastructure.S3/S3Options.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace FileService.Infrastructure.S3;
+﻿namespace FileService.Infrastructure.S3;
 
 public record S3Options
 {
